mic check 1..2 with a memory leak or 2

+++ minisound_ffi/lib/minisound_ffi.dart
@@ -306,7 +306,7 @@ class FfiRecorder implements PlatformRecorder {
   bool get isRecording => _bindings.recorder_is_recording(_self);
 
   @override
-  Uint8List getBuffer(int framesToRead) {
+  Float32List getBuffer(int framesToRead) {
     final output = malloc<Float>(framesToRead);
     try {
       final framesRead =
